mongoDB interface
gigantic push.... you might just say its MONGO hehe... so i've improved
all the finviz scripts... it is now automatic enough such that a cron
job could be run at the closing bell every day... working on getting
the mongo to a server where we can all share it...

insane speed improvements as well... because i've discovered threading
in python, we can scrape 1000 tickers at a time (yes you read that
right - 1000) so the scrape takes a total of 3 minutes. calculation of
mean metrics takes about 20 seconds, and back up to mongodb takes about
10 seconds. its really impressive... both python and mongo surpassed my
expectations today

devon: it would be nice to have the ticker script fixed as well as the
metric one i mentioned to you

ben: could we get some ideas for "learning" the market, or even just
some analysis tools? if i start scraping every day, we'll soon be
buried in data if we don't know what to do with

happy halloween!

#foreveralone